Delete chart button re-added. Publishers now add and remove instantly. Barchart and timelapse both updated and refactored.
