+++ app/components/chart-panel/donut.js
@@ -1,16 +1,8 @@
 import React from 'react'
 import Fluxxor from 'fluxxor'
-import KeywordList from './chart-menu/keywordlist'
-import AddKeyword from './chart-menu/addkeyword'
-import PublisherList from './chart-menu/publisherlist'
-import AddPublisher from './chart-menu/addpublisher'
-import Slider from './slider'
-import ChartTitle from './charttitle'
-import { Button } from 'react-bootstrap'
 
 var DonutChart = require("react-chartjs").Doughnut;
 var FluxMixin = Fluxxor.FluxMixin(React);
-var StoreWatchMixin = Fluxxor.StoreWatchMixin;
 
 var options = {
     //Boolean - Whether we should show a stroke on each segment
@@ -42,55 +34,23 @@ var options = {
 };
 
 export default React.createClass({
+
   mixins: [FluxMixin],
+
   componentDidMount: function() {
-    this.getFlux().actions.loadChartData(this.props.chartParams.chartID);
-  },
-  addKeyword: function(keyword){
-    this.getFlux().actions.addKeyword(this.props.chartParams.chartID, keyword);
-  },
-  addPublisher: function(publisher){
-    this.getFlux().actions.addPublisher(this.props.chartParams.chartID, publisher);
-  },
-  handleDeleteChart: function() {
-    this.getFlux().actions.deleteChart(this.props.chartParams.chartID);
+    this.getFlux().actions.loadChartData(this.props.chartID);
   },
 
   render: function() {
+
     var chart;
-    var deleteChartButton = "";
 
-    if (this.props.chartListLength > 1) {
-      deleteChartButton = (
-        <Button onClick={this.handleDeleteChart} className="delete" bsStyle="danger">Delete Chart</Button>
-      )
-    }
-
-    if(this.props.chartParams.snapShot){
+    if(this.props.data){
       chart = (
-        <div className="chart-container">
-          <div className="chart-label-y">
-          </div>
-          <div className="chart-main">
-          <ChartTitle title={this.props.chartParams.title} chartID={this.props.chartParams.chartID} />
-            <DonutChart className="chart"
-                      data={this.props.chartParams.snapShot}
-                      options={options}/>
-            <Slider chartID={this.props.chartParams.chartID} dates={this.props.allDates} startDate={this.props.chartParams.startDate} endDate={this.props.chartParams.endDate}/>
-          </div>
-          <div className="chart-menu">
-            <h5>Keywords</h5>
-            <AddKeyword chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} />
-            <ActiveKeywordList chartID={this.props.chartParams.chartID} className={'keyword-list'} list={this.props.chartParams.keywords} legend={true} />
-            <h5>Publishers</h5>
-              <ActivePublisherList chartID={this.props.chartParams.chartID}
-                                   className={'publisher-list'}
-                                   list={this.props.publisherList}
-                                   activelist={this.props.chartParams.publishers}/>
-              <AddPublisher chartID={this.props.chartParams.chartID} list={this.props.publisherList} activelist={this.props.chartParams.publishers} />
-              {deleteChartButton}
-          </div>
-        </div>
+        <DonutChart className="chart"
+                    data={this.props.data}
+                    options={options}
+        />
       )
     } else {
       chart = (<div className="loading"><img src="/img/loading.gif" /></div>)
@@ -102,8 +62,3 @@ export default React.createClass({
     )
   }
 })
-
-// <ActivePublisherList chartID={this.props.chartParams.chartID} className={'publisher-list'} list={this.props.publisherList} activelist={this.props.chartParams.publishers} />
-
-// redraw={this.props.chartParams.shouldRedraw}
-// // <Slider chartID={this.props.chartParams.chartID} dates={this.props.allDates} startDate={this.props.chartParams.startDate} endDate={this.props.chartParams.endDate}/>
